backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
yt-dlp>=2024.3.10
httpx==0.28.1
python-multipart==0.0.9
spotipy==2.23.0
python-dotenv==1.0.1
firebase-admin==6.5.0
fzmovies-api>=0.1.5
pydantic>=2.0.0
beautifulsoup4==4.12.3
